requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
pydantic
pwdlib[argon2]
python-dotenv
sqlmodel
sqlalchemy
passlib[argon2]